Refactor the whenever schedule to use Rake (#104)
Alan found issues with using the whenever command runner on the dev and QA environments. Using Rake instead of the command runner should help resolve these issues.

While making this change, the cause of an intermittent error when testing the exporter was found and resolved.

+++ lib/defra_ruby/exporters/registration_export_service.rb
@@ -9,9 +9,10 @@ module DefraRuby
 
       def epr_export
         file_path = full_path(EPR_EXPORT_FILENAME)
+        File.new(file_path, "wb")
         write_to_file(RegistrationEprExportReport, file_path)
         save_to_s3(:epr, file_path)
-        File.delete(file_path)
+        File.delete(file_path) if File.exist? file_path
       end
 
       def write_to_file(report_class, file_path)
@@ -67,7 +68,6 @@ module DefraRuby
 
         s3.bucket(aws_config[export_type][:bucket])
       end
-
     end
   end
 end
